Tests/LibWeb/Text/expected/console/console-prototype.txt

0
Prototype of console prototype is Object.prototype
